Stricter checks

- Updated compile flags/warnings for debug and test targets
- Several minor fixes
- Compilation for testing and debugging now requires at least gcc13 or clang-15

+++ src/SW_netCDF_Output.c
@@ -1258,7 +1258,7 @@ variable
 */
 void SW_NCOUT_create_output_dimVar(
     char *name,
-    unsigned int size,
+    size_t size,
     int ncFileID,
     int *dimID,
     Bool hasConsistentSoilLayerDepths,
